requirements.txt
streamlit==1.49.1
torch==2.8.0
torchvision==0.23.0
Pillow==10.4.0
numpy==2.3.3
gdown==5.2.0
opencv-python-headless>=4.8.0
matplotlib>=3.7.0
